+++ CHANGELOG.md
@@ -5,6 +5,9 @@ All notable changes to the Julia extension will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
+### Fixed
+- Fixed a race condition where user REPL configuration (e.g. `auto_insert_closing_bracket`) set in `startup.jl` via `atreplinit` could be overwritten ([#4063](https://github.com/julia-vscode/julia-vscode/pull/4063))
+
 ## [1.196.0] - 2026-03-16
 ### Fixed
 - Code evaluation is now more robust against unexpected errors ([#4056](https://github.com/julia-vscode/julia-vscode/pull/4056))